backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
dapr-sdk
